% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{checkDate}
\alias{checkDate}
\title{Check the date format.}
\usage{
checkDate(date)
}
\description{
\code{checkDate} apply several test on the Date and stop if problem detected
}
\details{
Details
}
