% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{getGFId}
\alias{getGFId}
\title{Get the GreenFeed ID of an animal object.}
\usage{
getGFId(theObject)
}
\description{
\code{getGFID} returns the greenfeed ID of the animal object to the gfid value.
}
\details{
Details
}
