% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{getMoves}
\alias{getMoves}
\title{Get the moves of an animal object.}
\usage{
getMoves(theObject)
}
\description{
\code{getMoves} returns the moves of the animal object to the moves value.
}
\details{
Details
}
