% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{getState}
\alias{getState}
\title{Get the state of an animal object.}
\usage{
getState(theObject)
}
\description{
\code{getState} returns the state of the animal object to the race value.
}
\details{
Details
}
