% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{getStateInfo}
\alias{getStateInfo}
\title{Get the state info of an animal object.}
\usage{
getStateInfo(theObject)
}
\description{
\code{getStateInfo} returns the state info of the animal object to the stateInfo value.
}
\details{
Details
}
