% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{getWeighing}
\alias{getWeighing}
\title{Get the Weighing values of an animal object.}
\usage{
getWeighing(theObject)
}
\description{
\code{getWeighing} returns the weights of the animal object to the weighing value.
}
\details{
Details
}
