% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFVisitData.R
\name{mergeData}
\alias{mergeData}
\title{merge weight and gf visit data.}
\usage{
mergeData(weights.xts, gf_em.xts)
}
\description{
\code{mergeData} returns the visit data from an animal with the interpolated weights info added
}
\details{
Details
}
