% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{readGender}
\alias{readGender}
\title{Read the gender of an animal.}
\usage{
readGender(id, file)
}
\description{
\code{readGender} returns the gender of the animal with the given id, as written in the file.
}
\details{
Details
}
