% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{readGFId}
\alias{readGFId}
\title{Read the list of GF Ids.}
\usage{
readGFId(id, file)
}
\description{
\code{readGFId} returns the list of 'GF ids' of the animal with the given id, as written in the file.
}
\details{
Details
}
