% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFVisitData.R
\name{readGFVisitData}
\alias{readGFVisitData}
\title{Read the greenfeed visit data.}
\usage{
readGFVisitData(gf_rfids, gffile)
}
\description{
\code{readGFVisitData} returns the visit data from an animal with RFID gf_rfid in a given greenfeed csv file
}
\details{
Details
}
