% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{readMoves}
\alias{readMoves}
\title{Read the moves of an animal.}
\usage{
readMoves(id, file)
}
\description{
\code{readMoves} returns the moves (xts) of the animal with the given id, as written in the file.
}
\details{
Details
}
