% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{readRace}
\alias{readRace}
\title{Read the race of an animal.}
\usage{
readRace(id, file)
}
\description{
\code{readRace} returns the race of the animal with the given id, as written in the file.
}
\details{
Details
}
