% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{readState}
\alias{readState}
\title{Read the state of an animal.}
\usage{
readState(id, file)
}
\description{
\code{readState} returns the state of the animal with the given id, as written in the file.
}
\details{
Details
}
