% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFromJournalsTools.R
\name{readWeighing}
\alias{readWeighing}
\title{Read the weighing of an animal.}
\usage{
readWeighing(id, file)
}
\description{
\code{readWeighing} returns the weighing values (xts) of the animal with the given id, as written in the file.
}
\details{
Details
}
