% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{setGFId}
\alias{setGFId}
\title{Set the GreenFeed ID of an animal object.}
\usage{
setGFId(theObject, gfid)
}
\description{
\code{setGFID} sets the greenfeed ID of the animal object to the gfid value.
}
\details{
Details
}
