% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{setMoves}
\alias{setMoves}
\title{Set the moves of an animal object.}
\usage{
setMoves(theObject, moves)
}
\description{
\code{setMoves} sets the moves of the animal object to the moves value.
}
\details{
Details
}
