% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{setRace}
\alias{setRace}
\title{Set the race of an animal object.}
\usage{
setRace(theObject, race)
}
\description{
\code{setRace} sets the race of the animal object to the race value.
}
\details{
Details
}
