% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{setState}
\alias{setState}
\title{Set the state of an animal object.}
\usage{
setState(theObject, state)
}
\description{
\code{setState} sets the state of the animal object to the race value.
}
\details{
Details
}
