% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{setStateInfo}
\alias{setStateInfo}
\title{Set the state info of an animal object.}
\usage{
setStateInfo(theObject, stateInfo)
}
\description{
\code{setStateInfo} sets the state info of the animal object to the stateInfo value.
}
\details{
Details
}
