% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{setWeighing}
\alias{setWeighing}
\title{Set the Weighing values of an animal object.}
\usage{
setWeighing(theObject, weighing)
}
\description{
\code{setWeighing} sets the weights of the animal object to the weighing value.
}
\details{
Details
}
